Create Frequency analysis of product reviews 2
Frequency analysis of product reviews
You now have access to a larger dataset of TechZone product reviews. Just like before, you've preprocessed and transformed the reviews into a BoW representation X. Your task now is to analyze the word frequencies and identify the most common terms in the dataset.

To help with the analysis, a helper function called get_top_ten() is provided. It takes in a list of words and their corresponding counts, and returns the 10 most frequent words and their counts.

Instructions 2/2

Modify the preprocess() function to remove stop_words as well.
Run the code to observe the new outcomes.